+++ packages/core/vitest/devices/closure.test.ts
@@ -592,10 +592,12 @@ describe('Matterbridge ' + NAME, () => {
     venetianBlind = new Closure('Venetian Blind Test Device', 'CL654321', {
       tagList: [getSemtag(ClosureTag.Covering), getSemtag(ClosureCoveringTag.Venetian)],
     });
-    venetianBlind.addPanel('Lift', [getSemtag(ClosurePanelTag.Lift)], 'lift');
+    venetianBlind.addPanel('Lift', [getSemtag(ClosurePanelTag.Lift)], 'lift', { motionLatching: true, speed: true });
     venetianBlind.addPanel('Tilt', [getSemtag(ClosurePanelTag.Tilt)], 'tilt', {
       resolution: 2,
       stepValue: 100,
+      motionLatching: true,
+      speed: true,
       latchControlModes: { remoteLatching: false, remoteUnlatching: true },
     });